+++ src/__tests__/cli-grammar.test.ts
@@ -197,4 +197,11 @@ test('settings grammar owns positional parsing for CLI commands', () => {
   assert.equal(clearAppState.setting, 'clear-app-state');
   assert.equal(clearAppState.state, 'clear');
   assert.equal(clearAppState.app, 'com.example.app');
+
+  const resetKeychain = readInputFromCli('settings', ['reset-keychain', 'clear'], {
+    ...BASE_FLAGS,
+    platform: 'ios',
+  });
+  assert.equal(resetKeychain.setting, 'reset-keychain');
+  assert.equal(resetKeychain.state, 'clear');
 });
