+++ src/daemon/handlers/__tests__/snapshot-handler.test.ts
@@ -20,7 +20,6 @@ import type { CaptureSnapshotResult } from '@agent-device/contracts/client';
 import { buildNodes } from '../../../__tests__/test-utils/snapshot-builders.ts';
 import {
   fixtureScreenshotCaptures,
-  fixtureSettingsMutations,
   resetSnapshotRuntimeFixture,
   snapshotRuntimeFixture,
 } from '../../__tests__/snapshot-runtime-fixture.ts';
@@ -34,7 +33,6 @@ import {
   inboxRow,
   iosSimulatorDevice,
   locationRequiredCapture,
-  macOsDevice,
   makeAndroidFreshnessSession,
   makeProviderRuntimeOwning,
   makeSession,
@@ -1439,125 +1437,6 @@ test('wait timeout without readable capture does not inspect the current surface
     expect(response.error.details?.readableCaptures).toBe(0);
   }
   expect(legacyDispatchCapture).not.toHaveBeenCalled();
-});
-
-test('settings rejects unsupported iOS physical devices', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'ios-device';
-  sessionStore.set(
-    sessionName,
-    makeSession(sessionName, {
-      platform: 'apple',
-      id: 'ios-device-1',
-      name: 'My iPhone',
-      kind: 'device',
-      booted: true,
-    }),
-  );
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'settings', { positionals: ['wifi', 'on'] }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response).toBeTruthy();
-  expect(response?.ok).toBe(false);
-  if (response && !response.ok) {
-    expect(response.error.code).toBe('UNSUPPORTED_OPERATION');
-    expect(response.error.message).toMatch(/settings is not supported/i);
-  }
-});
-
-test('settings clear-app-state dispatches explicit app id without an active app session', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'ios-clear-state';
-  sessionStore.set(sessionName, makeSession(sessionName, iosSimulatorDevice));
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'settings', {
-      positionals: ['clear-app-state', 'org.reactnavigation.playground'],
-    }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  expect(fixtureSettingsMutations.at(-1)).toMatchObject({
-    setting: 'clear-app-state',
-    state: 'clear',
-    appBundleId: 'org.reactnavigation.playground',
-  });
-});
-
-test('settings clear-app-state rejects missing app id when no app session is bound', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'ios-clear-state-missing-app';
-  sessionStore.set(sessionName, makeSession(sessionName, iosSimulatorDevice));
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'settings', { positionals: ['clear-app-state'] }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(false);
-  if (response?.ok === false) {
-    expect(response.error.code).toBe('INVALID_ARGS');
-    expect(response.error.message).toMatch(/requires an app id/i);
-  }
-  expect(fixtureSettingsMutations).toHaveLength(0);
-});
-
-test('settings usage hint documents canonical faceid states', async () => {
-  const sessionStore = makeSessionStore();
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest('default', 'settings'),
-    sessionName: 'default',
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response).toBeTruthy();
-  expect(response?.ok).toBe(false);
-  if (response && !response.ok) {
-    expect(response.error.code).toBe('INVALID_ARGS');
-    expect(response.error.message).toMatch(/appearance <light\|dark\|toggle>/);
-    expect(response.error.message).toMatch(/match\|nonmatch\|enroll\|unenroll/);
-    expect(response.error.message).toMatch(/grant\|deny\|reset/);
-    expect(response.error.message).not.toMatch(/validate\|unvalidate/);
-  }
-});
-
-test('settings on macOS rejects wifi before dispatch with explicit subset guidance', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'macos-settings-wifi';
-  sessionStore.set(sessionName, makeSession(sessionName, macOsDevice));
-
-  const response = await handleSnapshotCommands({
-    req: snapshotRequest(sessionName, 'settings', { positionals: ['wifi', 'on'] }),
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response).toBeTruthy();
-  expect(response?.ok).toBe(false);
-  expect(fixtureSettingsMutations).toHaveLength(0);
-  if (response && !response.ok) {
-    expect(response.error.code).toBe('INVALID_ARGS');
-    expect(response.error.message).toMatch(/Unsupported macOS setting: wifi/i);
-    expect(response.error.message).toMatch(/appearance <light\|dark\|toggle>/);
-    expect(response.error.message).toMatch(
-      /permission <grant\|reset> <accessibility\|screen-recording\|input-monitoring>/,
-    );
-    expect(response.error.message).toMatch(
-      /wifi\|airplane\|location\|animations remain unsupported on macOS/i,
-    );
-  }
 });
 
 test('diff rejects unsupported kind', async () => {
